Projects/PyCharm/Scripts/CPU_GPU_AlexNet_Performance/requirements.txt
torch
Pillow
torchvision
tqdm
